refactor project: move Wikimedia Kafka Producer Microservice to separate module
