+++ Wikimedia-Pet-Project/Recentchange-Producer-Microservice/src/main/java/yevhent/project/wikimedia/ApplicationProperty.java
@@ -1,4 +1,4 @@
-package yevhent.demo.kafka;
+package yevhent.project.wikimedia;
 
 public class ApplicationProperty {
